+++ cli/test/build-sandboxes.spec.ts
@@ -1,15 +1,14 @@
-import {} from 'jest';
 import { buildSandboxFileContents, findSandboxes, SandboxFileInformation, slash } from '../src/build-sandboxes';
 
 describe('findSandboxes', () => {
     let sandboxes;
 
     beforeEach(() => {
-        sandboxes = findSandboxes('./cli/test/files/sandboxes/');
+        sandboxes = findSandboxes(['./cli/test/files/sandboxes/', './cli/test/files/sandboxes-2/']);
     });
 
     it('should find sandbox files and assemble multiple sandboxes', () => {
-        expect(sandboxes.length).toBe(6);
+        expect(sandboxes.length).toBe(12);
     });
 
     it('should create a key that contains the path to the sandbox', () => {
@@ -98,50 +97,52 @@ describe('buildSandboxFileContents', () => {
         sandboxes = [
             {
                 key: 'cli/test/files/sandboxes/example1.sandbox',
+                srcPath: 'home/',
                 searchKey: 'ExampleComponent',
                 name: 'ExampleComponent',
                 label: '',
                 scenarioMenuItems: [
                     {
                         key: 1,
-                        description: 'Default'
+                        description: 'Default',
                     },
                     {
                         key: 2,
-                        description: 'With Wrapper'
-                    }
-                ]
+                        description: 'With Wrapper',
+                    },
+                ],
             },
             {
                 key: 'cli/test/files/sandboxes/other.sandbox',
+                srcPath: 'other-home/',
                 searchKey: 'OtherComponent',
                 name: 'OtherComponent',
                 label: '',
                 scenarioMenuItems: [
                     {
                         key: 1,
-                        description: 'Default'
-                    }
-                ]
-            }
+                        description: 'Default',
+                    },
+                ],
+            },
         ];
     });
 
     describe('exported functions', () => {
         it('should return a function, getSandboxMenuItems, with the JSON string contents of the sandboxes', () => {
-            const fileContents = buildSandboxFileContents(sandboxes, 'home/', 'lazy');
+            const fileContents = buildSandboxFileContents(sandboxes, 'lazy');
             expect(fileContents).toContain(JSON.stringify(sandboxes));
         });
 
         it('should return a function, getSandbox, that includes imports to sandbox files', () => {
-            const fileContents = buildSandboxFileContents(sandboxes, 'home/', 'lazy');
+            const fileContents = buildSandboxFileContents(sandboxes, 'lazy');
             expect(fileContents).toContain('import( /* webpackMode: "lazy" */ \'home/cli/test/files/sandboxes/example1.sandbox\')');
-            expect(fileContents).toContain('import( /* webpackMode: "lazy" */ \'home/cli/test/files/sandboxes/other.sandbox\')');
+            expect(fileContents).toContain('import( /* webpackMode: "lazy" */ \'other-home/cli/test/files/sandboxes/other.sandbox\')');
         });
 
 
         it('should include exports for both sandbox-fetching functions', () => {
-            const fileContents = buildSandboxFileContents(sandboxes, 'home/', 'eager');
+            const fileContents = buildSandboxFileContents(sandboxes, 'eager');
             expect(fileContents).toContain('exports.getSandboxMenuItems = getSandboxMenuItems;');
             expect(fileContents).toContain('exports.getSandbox = getSandbox;');
         });
@@ -149,7 +150,7 @@ describe('buildSandboxFileContents', () => {
 
     describe('getSandbox', () => {
         it('should return the serialized sandbox', () => {
-            const fileContents = buildSandboxFileContents(sandboxes, 'home/', 'lazy');
+            const fileContents = buildSandboxFileContents(sandboxes, 'lazy');
             expect(fileContents).toContain('return _.default.serialize(\'cli/test/files/sandboxes/example1.sandbox\')');
             expect(fileContents).toContain('return _.default.serialize(\'cli/test/files/sandboxes/other.sandbox\')');
         });
@@ -157,22 +158,22 @@ describe('buildSandboxFileContents', () => {
 
     describe('webpack strategy', () => {
         it('should use lazy webpack resolution strategy if lazy is provided as a parameter', () => {
-            const fileContents = buildSandboxFileContents(sandboxes, 'home/', 'lazy');
+            const fileContents = buildSandboxFileContents(sandboxes, 'lazy');
             expect(fileContents).toContain('import( /* webpackMode: "lazy" */ \'home/cli/test/files/sandboxes/example1.sandbox\')');
             expect(fileContents).not.toContain('eager');
         });
 
         it('should use eager webpack resolution strategy if lazy is provided as a parameter', () => {
-            const fileContents = buildSandboxFileContents(sandboxes, 'home/', 'eager');
+            const fileContents = buildSandboxFileContents(sandboxes, 'eager');
             expect(fileContents).toContain('import( /* webpackMode: "eager" */ \'home/cli/test/files/sandboxes/example1.sandbox\')');
             expect(fileContents).not.toContain('lazy');
         });
     });
-});
 
-describe('slash', () => {
-    it('should convert Windows-style path to unix-style', () => {
-        const windowsPath = 'c:\\etc\\';
-        expect(slash(windowsPath)).toBe('c:/etc/');
+    describe('slash', () => {
+        it('should convert Windows-style path to unix-style', () => {
+            const windowsPath = 'c:\\etc\\';
+            expect(slash(windowsPath)).toBe('c:/etc/');
+        });
     });
 });